Updated GroupCache to no longer return groups passed into the similar methods. This increases performance by avoiding checking distance to the source cluster. Obviously, the source cluster is close and similar to itself, but we want to calculate distance to different groups or clusters.
Added a profile to only run visual tests. Visual tests are tests that have no assertions but produce an image that can be reviewed. These can be expanded later to assert correctness.

Added some image utilities to make it easy to convert from and to 2D images.

Created an initial visual test.